feat: Goal Screen

+++ domain/src/main/java/me/abhigya/bourbon/domain/entities/User.kt
@@ -22,7 +22,7 @@ value class Centimeters(val value: Int = 0)
 
 @Serializable
 @JvmInline
-value class Kilograms(val value: Float = 0f)
+value class Kilograms(val value: Int = 0)
 
 enum class Gender {
     Male,
